# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PaperKitchen API",
        "description": "API for paperkitchen.ca",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://paperkitchen.adaptable.app/api",
            "description": "Production server for PaperKitchen"
        }
    ],
...
    "paths": {
...
        "/grocery-lists": {
            "get": {
                "tags": ["Grocery Lists"],
                "summary": "Get all grocery lists for the current user",
                "description": "Get all grocery lists for the current user",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [{ "ApiKeyAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Array of grocery lists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroceryList"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error getting grocery lists"
                    }
                }
            },
            "post": {
                "tags": ["Grocery Lists"],
                "summary": "Create a new grocery list",
                "description": "Create a new grocery list",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGroceryList"
                            }
                        }
                    }
                },
                "security": [{ "ApiKeyAuth": [] }],
                "responses": {
                    "201": {
                        "description": "Grocery list created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroceryList"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating grocery list"
                    }
                }
            }
        },
...
        "/recipes": {
            "get": {
                "tags": ["Recipes"],
                "summary": "Get all recipes",
                "description": "Retrieve a list of recipes based on query parameters.",
                "security": [{ "ApiKeyAuth": [] }],
                "parameters": [
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "pageSize",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recipes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Recipe"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "tags": ["Recipes"],
                "summary": "Create a recipe",
                "description": "Create a new recipe with the given details.",
                "security": [{ "ApiKeyAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecipe"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Recipe created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Recipe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "description": "Error message"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating recipe"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GroceryList": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer" },
                    "title": { "type": "string" },
                    "color": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" },
                    "ingredients": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Ingredient" }
                    }
                }
            },
            "CreateGroceryList": {
                "type": "object",
                "required": ["title", "color"],
                "properties": {
                    "title": { "type": "string" },
                    "color": { "type": "string" }
                }
            },
...
            "Ingredient": {
                "type": "object",
                "properties": {
                    "id": { "type": "string" },
                    "name": { "type": "string" },
                    "amount": { "type": "string" },
                    "unit": { "type": "string" }
                }
            },
            "Recipe": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer" },
                    "ownerId": { "type": "integer" },
                    "title": { "type": "string" },
                    "serves": { "type": "integer" },
                    "cookingTime": { "type": "integer" },
                    "cost": { "type": "integer" },
                    "favourite": { "type": "boolean" },
                    "color": { "type": "string" },
                    "type": { "type": "string" },
                    "createdAt": { "type": "string", "format": "date-time" },
                    "updatedAt": { "type": "string", "format": "date-time" },
                    "instructions": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Instruction" }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/Ingredient" }
                    }
                }
            },
            "CreateRecipe": {
                "type": "object",
                "properties": {
                    "title": { "type": "string" },
                    "color": { "type": "string" }
                }
            },
...
            "Instruction": {
                "type": "object",
                "properties": {
                    "id": { "type": "string" },
                    "text": { "type": "string" }
                }
            },
...
        }
    }
}